pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["abhi2596 <abhijeetchintakunta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.33.0"
llama-index = "^0.10.30"
llama-index-core = "^0.10.30"
llama-index-readers-file = "^0.1.19"
beautifulsoup4 = "^4.12.3"
html5lib = "^1.1"
unifyai = "^0.8.2"
pypdf2 = "^3.0.1"
llama-index-llms-openai-like = "^0.1.3"
fitz = "^0.0.1.dev2"
pymupdf = "^1.24.2"
sentence-transformers = "^2.2.2"
llama-index-embeddings-huggingface = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
